REF: Use topic and subscription context managers in more places

+++ octue/resources/service.py
@@ -1,9 +1,7 @@
 import json
 import logging
-import sys
 import time
 from concurrent.futures import TimeoutError
-from signal import SIGINT, signal
 import google.api_core.exceptions
 from google.cloud import pubsub_v1
 from icecream import ic
@@ -18,8 +16,9 @@ GCP_PROJECT = "octue-amy"
 
 
 class Topic:
-    def __init__(self, name):
+    def __init__(self, name, delete_on_exit=True):
         self.name = name
+        self._delete_on_exit = delete_on_exit
         self._publisher = pubsub_v1.PublisherClient()
         self.path = self._publisher.topic_path(GCP_PROJECT, self.name)
 
@@ -27,10 +26,22 @@ class Topic:
         return self
 
     def __exit__(self, exc_type, exc_val, exc_tb):
-        self._publisher.delete_topic(topic=self.path)
+        if self._delete_on_exit:
+            self._publisher.delete_topic(topic=self.path)
 
-    def create(self):
-        self._publisher.create_topic(name=self.path)
+    def create(self, allow_existing=False):
+        if not allow_existing:
+            self._publisher.create_topic(name=self.path)
+
+        else:
+            try:
+                self._publisher.create_topic(name=self.path)
+            except google.api_core.exceptions.AlreadyExists:
+                pass
+
+    def publish(self, data, attributes=None):
+        attributes = attributes or {}
+        self._publisher.publish(self.path, data, **attributes)
 
 
 class Subscription:
@@ -41,21 +52,24 @@ class Subscription:
         self.path = self.subscriber.subscription_path(GCP_PROJECT, self.name)
 
     def __enter__(self):
-        signal(SIGINT, self._sigint_handler)
         return self
 
     def __exit__(self, exc_type, exc_val, exc_tb):
         self.subscriber.delete_subscription(subscription=self.path)
+        self.subscriber.close()
 
-    def create(self):
-        self.subscriber.create_subscription(topic=self.topic.path, name=self.path)
+    def create(self, allow_existing=False):
+        if not allow_existing:
+            self.subscriber.create_subscription(topic=self.topic.path, name=self.path)
+
+        else:
+            try:
+                self.subscriber.create_subscription(topic=self.topic.path, name=self.path)
+            except google.api_core.exceptions.AlreadyExists:
+                pass
 
     def subscribe(self, callback=None):
         return self.subscriber.subscribe(self.path, callback=callback)
-
-    def _sigint_handler(self, signal_received, frame):
-        self.__exit__(None, None, None)
-        sys.exit(0)
 
 
 class Service:
@@ -73,6 +87,7 @@ class Service:
             ic(topic.path)
 
             with Subscription(name=self.name, topic=topic) as subscription:
+
                 subscription.create()
                 ic(subscription.path)
 
@@ -81,94 +96,73 @@ class Service:
                     question.ack()
 
                 streaming_pull_future = subscription.subscribe(callback=question_callback)
+                start_time = time.perf_counter()
 
-                with subscription.subscriber:
-                    start_time = time.perf_counter()
+                while True:
+                    if self._time_is_up(start_time, timeout):
+                        return
 
-                    while True:
-                        if self._time_is_up(start_time, timeout):
-                            return
+                    try:
+                        ic("Server waiting for questions...")
+                        streaming_pull_future.result(timeout=20)
+                    except Exception:
+                        # streaming_pull_future.cancel()
+                        pass
 
-                        try:
-                            ic("Server waiting for questions...")
-                            streaming_pull_future.result(timeout=20)
-                        except Exception:
-                            # streaming_pull_future.cancel()
-                            pass
+                    try:
+                        raw_question = vars(self).pop("_question")
+                    except KeyError:
+                        continue
 
-                        try:
-                            raw_question = vars(self).pop("_question")
-                        except KeyError:
-                            continue
+                    ic(f"Server got question {raw_question.data}.")
+                    question = json.loads(raw_question.data.decode())  # noqa
 
-                        ic(f"Server got question {raw_question.data}.")
-                        question = json.loads(raw_question.data.decode())  # noqa
+                    # Insert processing of question here.
+                    #
+                    #
+                    #
 
-                        # Insert processing of question here.
-                        #
-                        #
-                        #
+                    output_values = {}
+                    self.respond(output_values)
 
-                        output_values = {}
-                        self.respond(output_values)
-
-                        if exit_after_first_response:
-                            return
+                    if exit_after_first_response:
+                        return
 
     def respond(self, output_values):
-        publisher = pubsub_v1.PublisherClient()
-        topic_name = publisher.topic_path(GCP_PROJECT, f"{self.name}-response")
-
-        try:
-            publisher.create_topic(name=topic_name)
-        except google.api_core.exceptions.AlreadyExists:
-            pass
-
-        ic(f"Server responding on topic {topic_name}")
-        publisher.publish(topic_name, json.dumps(output_values).encode())
-        ic(f"Server responded on topic {topic_name}")
+        with Topic(name=f"{self.name}-response", delete_on_exit=False) as topic:
+            topic.create(allow_existing=True)
+            topic.publish(json.dumps(output_values).encode())
+            ic(f"Server responded on topic {topic.path}")
 
     def ask(self, service_name, input_values, input_manifest=None):
-        publisher = pubsub_v1.PublisherClient()
-        topic_name = publisher.topic_path(GCP_PROJECT, service_name)
-        publisher.publish(topic_name, json.dumps(input_values).encode())
-        ic(topic_name)
+        with Topic(name=service_name, delete_on_exit=False) as topic:
+            topic.publish(json.dumps(input_values).encode())
+            ic(topic.path)
 
     def wait_for_response(self, service_name, timeout=20):
-        publisher = pubsub_v1.PublisherClient()
-        topic_name = publisher.topic_path(GCP_PROJECT, f"{service_name}-response")
+        with Topic(name=f"{service_name}-response") as topic:
+            topic.create(allow_existing=True)
 
-        try:
-            publisher.create_topic(name=topic_name)
-        except google.api_core.exceptions.AlreadyExists:
-            pass
+            with Subscription(name=f"{service_name}-response", topic=topic) as subscription:
+                subscription.create()
 
-        subscriber = pubsub_v1.SubscriberClient()
-        subscription_name = subscriber.subscription_path(project=GCP_PROJECT, subscription=f"{service_name}-response")
-        subscriber.create_subscription(topic=topic_name, name=subscription_name)
+                def callback(response):
+                    self._response = response
+                    response.ack()
 
-        def callback(response):
-            self._response = response
-            response.ack()
+                ic(f"Asker waiting for response on {topic.path}")
+                future = subscription.subscribe(callback=callback)
 
-        ic(f"Asker waiting for response on {topic_name}")
-        future = subscriber.subscribe(subscription_name, callback)
+                try:
+                    future.result(timeout=timeout)
+                except TimeoutError:
+                    future.cancel()
 
-        with subscriber:
-            try:
-                future.result(timeout=timeout)
-            except TimeoutError:
-                future.cancel()
-
-            try:
-                print(self._response)
-                response = vars(self).pop("_response")
-            except KeyError:
-                pass
-
-            subscriber.delete_subscription(subscription=subscription_name)
-
-        publisher.delete_topic(topic=topic_name)
+                try:
+                    print(self._response)
+                    response = vars(self).pop("_response")
+                except KeyError:
+                    pass
 
         response = json.loads(response.data.decode())
         ic(f"Asker received response: {response}")
